backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==43.0.0
daphne==4.1.2
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.4.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
executing==2.0.1
gunicorn==22.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.0
ipython==8.26.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==71.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4.post2
